Validacion de Datos de Alumno API, y pruebas

+++ app/Http/Controllers/API/AlumnoController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use App\Models\Alumno;
 use App\Http\Resources\AlumnoCollection;
 use App\Http\Resources\AlumnoResource;
+use Validator;
 
 class AlumnoController extends BaseController
 {
@@ -30,9 +31,25 @@ class AlumnoController extends BaseController
      */
     public function store(Request $request)
     {
-        $data = request()->all();
-        $alumno = Alumno::create($data);
-        return new AlumnoResource($alumno);
+        $sizeCui = 8;
+        $rules =  [
+            'nombre'=>'required',
+            'apellido'=>'required',
+            'gmail'=>'required | email | unique:alumnos',
+            'cui'=>"required | size:{$sizeCui} | unique:alumnos",
+            'grupo_id'=>'nullable | exists:grupos,id'
+        ];
+        $input = $request->all();
+
+        $validator = Validator::make($input, $rules);
+        if($validator->fails()){
+            return $this->sendError('Validation Error.', $validator->errors());       
+        }
+
+        $alumno = Alumno::create($input)->refresh();
+        $data = new AlumnoResource($alumno);
+        
+        return $this->sendResponse($data, 'Alumno retrieved successfully', 201);
     }
 
     /**
@@ -44,6 +61,9 @@ class AlumnoController extends BaseController
     public function show($id)
     {
         $alumno = Alumno::find($id);
+        if($alumno == null ){
+            return $this->sendError('Alumno not found');
+        }
         $data = new AlumnoResource($alumno);
         return $this->sendResponse($data, 'Alumno retrieved successfully.');
     }
@@ -55,12 +75,30 @@ class AlumnoController extends BaseController
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    //**TODO */
     public function update(Request $request, $id)
     {
         $alumno = Alumno::find($id);
-        $inputs = $request->all();
-        $alumno->update($inputs);
+        if($alumno == null ){
+            return $this->sendError('Alumno not found');
+        }
+
+        $sizeCui = 8;
+        $rules =  [
+            'nombre'=>'sometimes | required',
+            'apellido'=>'sometimes | required',
+            'gmail'=>'sometimes | required | email |unique:alumnos,gmail,'.$alumno->id, 
+            'cui'=>"sometimes | required | size:{$sizeCui} | unique:alumnos".$alumno->id,
+            'grupo_id'=>'sometimes | nullable | exists:grupos,id'
+        ];
+        $input = $request->all();
+        $validator = Validator::make($input, $rules);
         
+        if($validator->fails()){
+            return $this->sendError('Validation Error.', $validator->errors());       
+        }
+      
+        $alumno->update($input);
         $data = new AlumnoResource($alumno);
 
         return $this->sendResponse($data, 'Alumno updated successfully.');
@@ -75,8 +113,12 @@ class AlumnoController extends BaseController
     public function destroy($id)
     {
         $alumno = Alumno::find($id);
-        $alumno->delete();
+        if($alumno == null ){
+            return $this->sendError('Alumno not found');
+        }
 
+        $alumno->delete();
+        
         $data = new AlumnoResource($alumno);
         return $this->sendResponse($data,'Alumno deleted successfully.');
     }
